phonefsod-dbus: delay showing the PIN dialog when phoneuid appears on the bus
If we immediately try to show the PIN dialog when we get notification
about phoneuid being on the bus that sometimes does not work - especially
on faster machines. We now delay the call to phoneuid for 2 seconds
which seems to be sufficient for phoneuid to finish
startup and initialization.

fixes trac #1478.

Signed-off-by: Klaus Kurzmann <mok@fluxnetz.de>